Separate model tests to separate tasks/directory.
refs #921
