+++ Gruntfile.js
@@ -237,6 +237,10 @@ var path           = require('path'),
                         'core/test/unit/**/export_spec.js',
                         'core/test/unit/**/import_spec.js'
                     ]
+                },
+
+                integration: {
+                    src: ['core/test/integration/**/model*_spec.js']
                 }
             },
 
@@ -866,9 +870,11 @@ var path           = require('path'),
 
         grunt.registerTask('test-unit', 'Run unit tests', ['clean:test', 'setTestEnv', 'loadConfig', 'express:test', 'mochacli:all']);
 
+        grunt.registerTask('test-integration', 'Run integration tests', ['clean:test', 'setTestEnv', 'loadConfig', 'express:test', 'mochacli:integration']);
+
         grunt.registerTask('test-functional', 'Run casperjs tests only', ['clean:test', 'setTestEnv', 'express:test', 'spawn-casperjs']);
 
-        grunt.registerTask('validate', 'Run tests and lint code', ['jslint', 'test-unit', 'test-functional']);
+        grunt.registerTask('validate', 'Run tests and lint code', ['jslint', 'test-unit', 'test-integration', 'test-functional']);
 
         grunt.registerTask('docs', 'Generate Docs', ['groc']);
