+++ core/test/integration/model_tags_spec.js
@@ -1,5 +1,5 @@
 /*globals describe, before, beforeEach, afterEach, it */
-var testUtils = require('./testUtils'),
+var testUtils = require('../unit/testUtils'),
     _ = require("underscore"),
     when = require('when'),
     sequence = require('when/sequence'),
